// components/home/type.ts
export interface HeroBannerConfig
{
    id: string;
    imageUrl: string;
    badge: string;
    title: string;
    description: string;
    cta: string;
    backgroundColorDark: string;
    textColorDark: string;
    textColorLight: string;
    backgroundColorLight: string
    badgeText: string
    badgeNavigationUrl: string
}